fix(server): limit maximum image size
The new maximum is 5000px for width and height. This seems to be a good
fit for all formats besides GIF for which generation of pictures that
big is still very slow (about 30 seconds on my machine).

The main concern I’m having about this issue is that this makes a DDOS
attack a thing you could do by hand.

I’m not sure how to handle this best. The options that come to mind are:
- dropping GIF support altogether
- special size limit for GIFs
- make GIF support configurable