chore(Docker): Switch to port 80 as exposed port
This is due to the fact that tools like Dokku use the EXPOSE directive
as an indicator which ports the container should be exposed publicly.

Since in such cases port 80 is probably the most common port the
Dockerfile is now configured to build the app to expose and listen on
that.

In other usecases like local development it is easy to just configure
another port mapping.